stress-ng: fix --permute with checksum verification enabled
Enabling --verify with the --permute option currently fails because
the Nth stress test does not map to the Nth checksum because the
checksums on ignore permutations aren't bumping the checksum pointer
so we get and mismatch on the mapping. Fix this by incrementing the
checksum pointe by the number of instances of ignored permutations.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>