stress-crypt: fix performance issues with crypt_r data
The re-entrant crypt_r data needs to be setup correctly on each
loop iteration to fix a performance issue. Also replace modulo
with bit-wise and operation to speed up orig_phrase setup since
the modulo is a power of two in size.

Kudos to Michael Larabel for reporting this issue.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>