stress-mutex: fix some potential race issues
Ensure all the pthread return values are initialized to -1 before
attempting to create threads so we can detect value pthreads to
join later. Set created to true when a pthread is created before
breaking from the creation loop when stress_continue is false.

Accumulate duration and count after the join for non-racy update
of stats.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>