core-mwc: re-work mwc*modn() using bit masking and removing modulo operation
Use a bitmask that is large enough for all the digits in the modulo value
to generate a randomized value that is wide enough to get enough random
bits but no so wide that the thresholding loop does not loop too many
times. This allows removal of the final modulo which makes the code faster
for processors with slow integer modulo math.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>