mmapfork: add --mmapfork-bytes option to specify mmap size
Default mmapfork region size is 'DRAM_SIZE / (instances * 32) / 2' for
each worker. This value can be quite large on systems with a lot of
memory. As a result, mmapfork on slow systems may fail to complete in
time, if a short timeout is specified.

Add '--mmapfork-bytes' option to specify mmapfork test region size.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>