stress-min-nanosleep: add stressor that measures minimum nanosleep delay times
This stressor exercises various power of 2 nanosleep delays (by default from 0
to 1024*1024) and measures min, max and mean times for the nanosleeps based on
an average of 10 calls per measurment round. One can specify the type of scheduler
to use via the --min-nanosleep-sched option and also the maximum sleep via
the --min-nanosleep-max option.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>