Add --sync-start option
It's useful to be able to spawn stressor instances and wait until
they are in a ready-to-run state and then synchronously start them
all together in the run stressing state.

Add a new --sync-start option to allow this. This change adds
some helper functions that wait for the sync start by SIGSTOP'ing
the processes re-starting them using SIGCONT.

For stressors that spawn their own processes there are helpers
to mmap and unmap a shared memory stress_pid_t pids array and
to link the pids together, and to start them in a synchronous
manner using the list of pids.

Closes https://github.com/ColinIanKing/stress-ng/issues/397

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>