stress-ng: stress_sync_start_cont_list: check against number of pids in list
Currently the ready count is being used to check for number of stressors
that were in the ready state, however, stressors can bail out before the
ready state, for example when resouces are not available in the earlt init
phase. Hence use the number of pids in the list to use as the ready count.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>