stress-bitops: Add new bitops stressor --bitops
Add a stressor that exercises some commonly used bit manipulation
operations with different ways to perform the computation. Many of these
are derived from the Standford bithacks and some from Hacker's Delight.

For most methods a niave implementation is used as a reference
computation and subsequent ways to perform the computation are the
more optimal branchless implementations.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>