test-lib*: reference the contents of the function arrays
Modern optimizing compilers can detect that the array of functions
is never referenced causing linking to succeed even when the functions
are not linked against a library that supports them. We need these
to fail on linking when testing for library supported functions, so
iterate over the contents to ensure they are referenced and that the
linker does it's work, hence detecting missing functions when the
library does not supply them.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>