stress-spinmem: new spin write/read parent/child stressor
The spinmem stressor exercises inter-process communication via
a shared memory page and heavy spinloop usage on two memory locations.
One bogo-op is 1000 memory transactions between the two processes.

One process is a writer that increments a value in memory slot 0 and
spin waits for the data to appear in memory slot 1. The  other process
spin  waits forthe data to change in slot 0 and copies the changed value
into memory slot 1.  The stessor benchmarks the time for the transactions
to occur between both processes. Note that the optimal number of
stressors is half the number  of online CPUs in a system. By default
32 bit write/reads are used.

A spinmem method option allows selection of 8bit, 16bit, 32bit and
64bit memory write/read sizes.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>