stress-opcode: set child signal handlers before stress_shared_readonly
In the case of an early SIGALRM occuring before the child signal handler
being set we hit SIGSEGV in the default SIGALRM handler because the
stress_shared_readonly call has made the global data read-only. Fix this
by moving the signal handler setting to occur before the call to
stress_shared_readonly. Also add a flag to indicate that the jmp_env
is set so that we can avoid doing jmps to via this jmp_env if it's
not yet set up.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>