core-lock: use a single mapped page for all shared locks
Currently each lock required a shared page to allow multiple processes
to act on a lock, and so many pages could be allocated for all the
locks in use.  Re-work the locks so that they are much smaller and
are get/put from a single shared page. This allows 512 locks per 4K
page to be in use. The magic is now used as a lock valid patter and
set to zero when a lock is free and is available when a new lock is
required. A special lock (the big lock, lock index 0) in the shared
page when a new lock is searched for in the lock get operation and
when an old lock is put back when it freed.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>