stress-lsm: move ctx and ctx_end to inner loop scope
Currently ctx is not being reset to the start of the buf on each
loop, so we can end up with it falling off the end of buf. Move
the declaration so it's inside of the loop to ensure it gets
reset to the start of buf on each loop iteration.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>