Makefile: improve machine detection and disable -fmodulo-sched for S390x builds
Turns out that some S390x gcc toolchains take forever to build vector code
with -fmodulo-sched, so disable this for s390x builds. To do this the
MACHINE arch has been improved to build a test program and use readelf
to inspect the arch string to get the target binary arch. This also fixes
some cross-compilation arch checks that used to determine the MACHINE
setting from the host arch (which is incorrect).

Closes: https://github.com/ColinIanKing/stress-ng/issues/417

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>