stress-dev: exercise /dev/cpu/0/msr ioctl reading of TSC
The msr device can read MSRs two ways. One by using the arcane
X86_IOC_RDMSR_REGS ioctl where one passes 8 x 32 bit values that
map onto various registers used before an rdmsr opcode with the
MSR specified in regs[1] that maps to the ECX for the rdmsr opcode.
In this test case we read MSR 0x00000010, namely the TSC. The
contents of EDX:EAX is written into regs[2]:regs[0]. The second
easier way is to seek to the required MSR and do a read.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>