stress-chattr: add mmap'd page modification to file, add metrics
chattr can make a page immutable hence any mmap'd pages on the file
being chattr'd can throw a BUS error (or SEGV) if the page is modified,
so add a test case for this. Also add a successfuly chattr flag setting
metric. Finally, don't bail out if chattr fails, some attributes are
not available depending on the owner or file system but this should not
be treated as fatal.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>