stress-touch: Fix file name handling in stress_touch_dir_clean().
On Haiku R1/beta5 the code fails to compile:

stress-touch.c: In function 'stress_touch_dir_clean':
stress-touch.c:174:48: error: invalid application of 'sizeof' to
  incomplete type 'char[]'
  174 |                 char filename[PATH_MAX + sizeof(d->d_name) + 1];
      |                                                ^

/system/develop/headers/posix/dirent.h indeed defines the structure
member as "char d_name[]" (unlike, for example, Linux).  The length of
the file name is supposed to fit into NAME_MAX, so use it instead of the
sizeof() and add a comment to clarify the buffer sizing.  While at it,
lose the trailing newline character in the resulting buffer.