core-helper: add root like detection for Cygwin
Fix as per suggestion of Christian Franke:

By default, Cygwin does never return 0 for uid or gid unless there are
custom /etc/passwd or /etc/group which remap this, see
https://cygwin.com/cygwin-ug-net/ntsec.html#ntsec-mapping-how

There is no dedicated root-like user on Windows. The Local Adminstrator
account is typically not used, the SYSTEM account only by Windows internally.
Any user may have admin privileges. Even if this is the case, processes are
typically started without admin privileges (non-elevated) by default (UAC).
As a consequence, the uid is not sufficient to check for admin privileges.

Kudos to Christian Franke for the main code behind this fix.

Closes https://github.com/ColinIanKing/stress-ng/issues/430

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>