stress-prio-inv: Fix incorrect check for PTHREAD_PRIO_{INHERIT|NONE|PROTECT}
The checks for PTHREAD_PRIO_* was based on the assumption that these are
macros, instead they are enums and so this check was always false. Fix this by
adding build time checks for these enums and defining HAVE_PTHREAD_PRIO_*
if the PTHREAD_PRIO_* enums (or macros on some systems) exist.

Closes: https://github.com/ColinIanKing/stress-ng/pull/442
Closes: https://github.com/ColinIanKing/stress-ng/pull/443

Fixes: f57cf368ce7e ("core-*, stress-*: use new table-driven option parsing")
Signed-off-by: Colin Ian King <colin.i.king@gmail.com>