core-affinity: add helpers to get/free array of useable cpus
It's useful to obtain an array of cpu ids of usable cpus. The
stress_get_usable_cpus() function will either return an array
of all available cpu id's if use_affinity is false or an array
of all available cpu ids that are in the global affinity cpu set
(as determined by the --taskset option.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>