stress-ng: re-work aggressive pid waiting to use stress_wait_pid
Make stress_wait_pid a little more generic with a waitpid flag and
re-work the aggressive pid waiting loop to use this. Add a reaped
flag to the stress_pid_t struct to indicate whether a process has
been reaped using wait* to avoid redundant re-waiting on a reaped
process.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>