core-asm-ret: add lpad instruction in return code on RISC-V
Control-flow Integrity (CFI) supports Zicfiss and Zicfilp
extensions that provide backward-edge and forward-edge
control flow integrity respectively.

Forward-edge control flow introduces a landing pad (LPAD)
instruction that must be placed at the program locations
that are valid targets of indirect jumps or calls.

Without lpad instruction in return code here, some stress
tests will fail due to CFI violation such as far-branch,
flushcache and icache, when kernel enables the CFI config
and the platform supports Zicfilp extension.

This commit adds lpad instruction with label zero, because
lpad instruction does not perform the label check when the
label is encoded as zero, then we don't need to set the x7
register on caller site.

The lpad instruction is encoded using the AUIPC major opcode
with rd=x0, the lpad instruction executes as a no-op when
Zicfilp is not active or when the extension is not implemented.

Signed-off-by: Zong Li <zong.li@sifive.com>