stress-easy-opcode: add a simple easy opcode stressor
The easy opcode stressor generates 64 pages of randomized
easy to execute opcodes. For CISC processors such as x86 the
opcodes are simple 1 byte instructions such as nop and cpu status
set/clr instructions to cause the decoding to be front-end bound
and to incur instruction cache load. For RISC processors the
opcodes are simple instructions such as nop and register moves
(such as move r0,r0) that do little. On smart emulators that
perform instruction removal this can lead to misleading opcode
execution rates.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>