stress-fpunch: don't shared fd on child processes when using lseek
Sharing an fd with child processes when using lseek is a bad idea
since lseek will set the offset and the child processes sharing the
fd will be using that offset. Instead, for cases where preadv/pwritev
are not available and lseek is being used each child process requires
it's own fd.

Kudos to Christian Franke for reporting this issue in a detailed
manner.

Closes: https://github.com/ColinIanKing/stress-ng/issues/500

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>