stress-ng: re-work stress_continue() to just check bogo-ops
Rather checking of a global flag is true, and check for bogo-ops
reaching the maximum, rework to just check on bogo-ops. The
underlying mechanism is that the max bogo-ops is zero'd when
it's time to stop. A run-for-ever setting sets the max bogo-ops
to 0xffffffffffffffff rather than the original zero which needed
special checking first.

Note that a global flag g_stress_continue_flag is used for
args-less stress_continue_flag(). This may be phased out in
future.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>