stress-procfs: increase $PID count to 5 processes
Currently only a small number of processes are being exercised with
the stressor. Increase this to a handful, this is still very limited
but processing many PIDs is exercising the same kind of procfs files
and is expensive, we want this stressor to focus on a wide range of
procfs entries include many non-PID related entries.

Closes https://github.com/ColinIanKing/stress-ng/issues/523

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>