core-bitops: remove optimization hint on inline functions
If the optimization level does not match the same level as the
function that's calling the mwc*modn functions then the code is
not inlined and a wrapper is created to call these functions. This
is suboptimal. The best strategy is to remove the optimization hint
so the inline functions get inlined and then optimized at the level
of the calling function.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>