stress-filename: add readdir() filename sanity checking
For a random subset of the generated files, check the readdir()
result after file creation:

* Report failure if readdir() does not return exactly one name (after
  ignoring . and ..), else ...
* report failure if the name differs from the created file and stat()
  fails or suggests that this is not the same file, else ...
* if the name differs but refers to the same file, print a related information
  once per run.

The latter is not an error because bijective en/decoding between API and
filesystem is not always possible. This is in particular the case on Cygwin
due to a non-trivial conversion between current codepage and UTF-16.

Kudos to Christian Franke for suggesting this improvement.

Closes: https://github.com/ColinIanKing/stress-ng/issues/545

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>