core-parse: fix percentage file size args based on --temp-file option
Currently the % file size is based on the --temp-file setting at the
time of the args being parsed, this is too early since --temp-file may
be set before or after the file size parsing. If --temp-file is set
after the file size parsing then the scaling is based on the default
--temp-file setting (which is ".") which is incorrect. Fix this by
adding TYPE_ID_UINT64_BYTES_FS_PERCENT and TYPE_ID_SIZE_T_BYTES_FS_PERCENT
type identifiers and stashing values that are file size percentages
as these types and evaluate the actual value when getting these values
when the stressors are being invoked so that any settings of --temp-path
are now known.

Add extra bounds checking for these sizes as these are only know once
the stessor is invoked. Finally, add some overflowed uint64_t and size_t
percentage value checking.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>