core-memory: replace stress_munmap_retry_enomem with stress_munmap_force
There are a couple of reasons why munmap can fail, the first with ENOMEM
because there is no memory left to split a mapping, the second is an
EINVAL when a munmap of a HUGETLB mapping region that was mapped with
less than the huge page size is unmapped. In the EINVAL case, scan
the process mappings and determine if the mapping was a HUGETLB, get
the mapping size and if it looks like a sane HUGETLB mapping then unmap
the mapping with the known HUGETLB size.

Rename stress_munmap_retry_enomem to stress_munmap_force as the function
now forces unmappings on various errors rather than just the ENOMEM
case.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>