stress-mmaprandom: add RB tree for randomized node selection
Add another RB tree to the mr_node_t that is sorted on rand_id, a
randomized id that is unique (lower bits based on mr_node index
in the mr_nodes array, upper bits are randomized). This allows fast
selection of a random mr_node (by pulling it from the head) and
re-ordering when users nodes are added to the tree.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>