stress-cpu: introduce matrixprod64 benchmark
This benchmark is a copy-paste of the matrix prod benchmark, but made to
be less dependent on the x87 instruction set, and use standard 64-bits
floats instead.

On arm64 (for example), matrixprod gives very bad performance because of
the 80-bits float emulation.

The ratio was computed between matrixprod and matrixprod64 on an Intel
Core i7-10850H.

matrixprod64 should also use double the data cache because of the
matrices sizes going from 128x128 (80-bits, but aligned to 16 bytes, so
256KiB) to 256x256 (64-bits, so 512KiB).