stress-enosys: improve enosys performance
Run enosys calls inside a child process that receives syscall numbers
from parent via a pipe. Rather than forking for each syscall, we keep
a child running that the handles syscalls pushed to it from the parent
via a simeple rpc mechanism. The child writes back the errno and number
of system calls made back to the parent over a second pipe. If the
syscall makes the child exit then we re-spawn the child. This is
improves overall syscall ENOSYS rates. Also simplify the syscall
arguments since we just need to exercise a ENOSYS call and varying
the call arguments does not matter. Add more randomness into the system
call number ordering generation in parent for wider system call
range coverage. Remove legacy int 80 call for x86 as this uses a
system call numbering that may be different from the 64 bit call
mechanism and tracking these adds more unncessary complexity to the
stressor.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>