stress-timermix: Add a new mixed timer and itimer stressor
This stressor runs various high resolution timers and itimers
concurrently in each stressor instance. It will attempt to auto-adjust
the high resolution timer frequency to maximize the signal generation
rate of the timers. This scheduler is designed to generate a
high amount of timer interrupts and signals across a range of different
timer types.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>