Makefile.machine: use OBJDUMP when available
When cross-compiling, CC is defined as the cross-C compiler. However, in
Makefile.machine, we use the host "objdump", resulting in the error:

objdump: can't disassemble for architecture UNKNOWN!

This commit introduces the use of OBJDUMP, with a default value to
"objdump". This allows defining OBJDUMP as objdump from the toolchain (as
done for example in Buildroot).

Signed-off-by: Jimmy Durand Wesolowski <jimmy.wesolowski@mobileye.com>