stress-ng: add --oom-no-child option to remove OOM child wrapper
Sometimes it's useful to run a potentially OOMing process without
the OOM handling parent/child wrapper. This just runs the stressor
as a normal process and not a child of a OOM detecting parent.

The downside is that OOM restarts are disabled, OOM accounting may
not work and bogo-op verification may fail if a process is OOM'd
while it's bumping a bogo-op counter. The upside is that we remove
the need for another process per stressor instance for OOMable
stressors such as the vm stressor.

Signed-off-by: Colin Ian King <coking@nvidia.com>