stress-varyload: add a new varying load stressor
The varyload stressor allows one to create varying loads by changing
the number of varyload instances that are runnable. The workloads are
the same as the workload stressor. The --varyload-type option allows
one to change the way the load is changed using triangle, saw-inc,
saw-dec, pulse and random load shaping. By default the load is changed
every 1000ms, but this can be changed using the --varyload-ms option.

Signed-off-by: Colin Ian King <coking@nvidia.com>