stress-ng: add new option --make-it-fail to enable error injection faults
Enable error injection faults on stressor processes (and their child
processes) by writing to /proc/self/make-it-fail.

The appropriate Linux kernel /sys/kernel/debug fault injection sysfs
interfaces need to be set inconjunction when using this option and
the kernel requires fault injection to be enabled. Needs to run
with root permission to write to /proc/self/make-it-fail.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>