stress-syscall: Fix blocking BSD system call issues
BSD can hang on copy_file_range if input file is output file, so
just write to /dev/null instead of to the src file. Also add
small timeout to complement the sched_yield calls as this helps
some BSD flavours block on pause calls that get signalled in the
signalling process is spinning on sending signals to the paused
child.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>