stress-pthread: remove alternative pthread stacks
Creating and joining thousands of pthreads is problematic with
alternative pthread stacks. POSIX already defines that the stack should
stay after creation. NetBSD pthreads implementation still requires
the stack to be around even after a pthread_join has occurred, so
this is problematic when freeing up a pthread's stack after a join.

Remove the alternative stacks to allow the code to be more portable.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>