stress-stream: reduce stream buffer size by factor of 4
There are 3 stream buffers that each are allocated the size of the L3
cache size. Currently these are scaled by a factor of 4 to ensure
that cache misses occur during the streaming. This is way too large,
just 3 buffers of the cache size is plenty, there is no need to
scale by 4. Remove this scaling.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>